Rename and use IBlockManagerBasic as IBlockManager and IChainManagerBasic as IChainManager
