+++ AElf.Contracts.Authorization.Tests/MockSetup.cs
@@ -53,8 +53,8 @@ namespace AElf.Contracts.Authorization.Tests
             var transactionManager = new TransactionManager(_dataStore, _logger);
             var transactionTraceManager = new TransactionTraceManager(_dataStore);
             _functionMetadataService = new FunctionMetadataService(_dataStore, _logger);
-            var chainManagerBasic = new ChainManagerBasic(_dataStore);
-            ChainService = new ChainService(chainManagerBasic, new BlockManagerBasic(_dataStore),
+            var chainManager = new ChainManager(_dataStore);
+            ChainService = new ChainService(chainManager, new BlockManager(_dataStore),
                 transactionManager, transactionTraceManager, _dataStore, StateStore);
             _smartContractRunnerFactory = new SmartContractRunnerFactory();
             var runner = new SmartContractRunner("../../../../AElf.Runtime.CSharp.Tests.TestContract/bin/Debug/netstandard2.0/");
@@ -68,7 +68,7 @@ namespace AElf.Contracts.Authorization.Tests
                 await Init();
             }).Unwrap().Wait();
             SmartContractService = new SmartContractService(SmartContractManager, _smartContractRunnerFactory, StateStore, _functionMetadataService);
-            ChainService = new ChainService(new ChainManagerBasic(_dataStore), new BlockManagerBasic(_dataStore), new TransactionManager(_dataStore), new TransactionTraceManager(_dataStore), _dataStore, StateStore);
+            ChainService = new ChainService(new ChainManager(_dataStore), new BlockManager(_dataStore), new TransactionManager(_dataStore), new TransactionTraceManager(_dataStore), _dataStore, StateStore);
         }
 
         private void NewStorage()
