+++ AElf.Miner.Rpc/Client/ClientManager.cs
@@ -31,7 +31,7 @@ namespace AElf.Miner.Rpc.Client
         private readonly ICrossChainInfo _crossChainInfo;
         private CertificateStore _certificateStore;
         private readonly ILogger _logger;
-        private readonly IChainManagerBasic _chainManagerBasic;
+        private readonly IChainManager _chainManager;
         private Dictionary<string, Uri> ChildChains => GrpcRemoteConfig.Instance.ChildChains;
         private CancellationTokenSource _tokenSourceToSideChain;
         private CancellationTokenSource _tokenSourceToParentChain;
@@ -42,10 +42,10 @@ namespace AElf.Miner.Rpc.Client
         /// </summary>
         private int WaitingIntervalInMillisecond => GrpcLocalConfig.Instance.WaitingIntervalInMillisecond;
 
-        public ClientManager(ILogger logger, IChainManagerBasic chainManagerBasic, ICrossChainInfo crossChainInfo)
+        public ClientManager(ILogger logger, IChainManager chainManager, ICrossChainInfo crossChainInfo)
         {
             _logger = logger;
-            _chainManagerBasic = chainManagerBasic;
+            _chainManager = chainManager;
             _crossChainInfo = crossChainInfo;
             GrpcRemoteConfig.ConfigChanged += GrpcRemoteConfigOnConfigChanged;
         }
@@ -97,7 +97,7 @@ namespace AElf.Miner.Rpc.Client
 
         private async Task<ulong> GetSideChainTargetHeight(Hash chainId)
         {
-            var height = await _chainManagerBasic.GetCurrentBlockHeightAsync(chainId);
+            var height = await _chainManager.GetCurrentBlockHeightAsync(chainId);
             return height == 0 ? GlobalConfig.GenesisBlockHeight : height + 1;
         }
         
@@ -262,7 +262,7 @@ namespace AElf.Miner.Rpc.Client
         /// <returns></returns>
         private async Task UpdateCrossChainInfo(IBlockInfo blockInfo)
         {
-            await _chainManagerBasic.UpdateCurrentBlockHeightAsync(blockInfo.ChainId, blockInfo.Height);
+            await _chainManager.UpdateCurrentBlockHeightAsync(blockInfo.ChainId, blockInfo.Height);
         }
 
         public bool TryGetSideChainBlockInfo(SideChainBlockInfo scb)
