+++ AElf.Miner.Tests/MinerLifetime.cs
@@ -352,8 +352,8 @@ namespace AElf.Kernel.Tests.Miner
                 Assert.Equal(1, result.IndexedBlockInfo.Count);
                 Assert.True(result.IndexedBlockInfo.Keys.Contains(GlobalConfig.GenesisBlockHeight));
                 //Assert.True(await manager.UpdateParentChainBlockInfo(result));
-                var chainManagerBasic = _mock.MockChainManager().Object;
-                await chainManagerBasic.UpdateCurrentBlockHeightAsync(result.ChainId, result.Height);
+                var chainManager = _mock.MockChainManager().Object;
+                await chainManager.UpdateCurrentBlockHeightAsync(result.ChainId, result.Height);
                 _mock.GetTimes++;
                 
                 Thread.Sleep(t);
@@ -363,7 +363,7 @@ namespace AElf.Kernel.Tests.Miner
                 Assert.Equal(1, result.IndexedBlockInfo.Count);
                 Assert.True(result.IndexedBlockInfo.Keys.Contains(GlobalConfig.GenesisBlockHeight + 1));
                 //Assert.True(await manager.UpdateParentChainBlockInfo(result));
-                await chainManagerBasic.UpdateCurrentBlockHeightAsync(result.ChainId, result.Height);
+                await chainManager.UpdateCurrentBlockHeightAsync(result.ChainId, result.Height);
                 _mock.GetTimes++;
 
                 Thread.Sleep(t);
