+++ AElf.Miner/Miner/Miner.cs
@@ -50,7 +50,7 @@ namespace AElf.Miner.Miner
 
         private Address _producerAddress;
         private ECKeyPair _keyPair;
-        private readonly IChainManagerBasic _chainManagerBasic;
+        private readonly IChainManager _chainManager;
         private readonly ConsensusDataProvider _consensusDataProvider;
 
         private IMinerConfig Config { get; }
@@ -64,7 +64,7 @@ namespace AElf.Miner.Miner
             ILogger logger, ClientManager clientManager,
             IBinaryMerkleTreeManager binaryMerkleTreeManager, ServerManager serverManager,
             IBlockValidationService blockValidationService, IChainContextService chainContextService
-            , IChainManagerBasic chainManagerBasic,IStateStore stateStore)
+            , IChainManager chainManager,IStateStore stateStore)
         {
             _txHub = txHub;
             _chainService = chainService;
