+++ AElf.Synchronization.Tests/ConfigureTestFramework.cs
@@ -32,8 +32,8 @@ namespace AElf.Synchronization.Tests
             builder.RegisterType<ChainContextService>().As<IChainContextService>().SingleInstance();
             builder.RegisterType<ChainService>().As<IChainService>().SingleInstance();
             builder.RegisterType<BlockSet>().As<IBlockSet>().SingleInstance();
-            builder.RegisterType<ChainManagerBasic>().As<IChainManagerBasic>().SingleInstance();
-            builder.RegisterType<BlockManagerBasic>().As<IBlockManagerBasic>().SingleInstance();
+            builder.RegisterType<ChainManager>().As<IChainManager>().SingleInstance();
+            builder.RegisterType<BlockManager>().As<IBlockManager>().SingleInstance();
             builder.RegisterType<TransactionManager>().As<ITransactionManager>().SingleInstance();
             builder.RegisterType<StateStore>().As<IStateStore>();
             builder.RegisterType<TxSignatureVerifier>().As<ITxSignatureVerifier>();
