+++ AElf.Synchronization.Tests/MockSetup.cs
@@ -30,7 +30,7 @@ namespace AElf.Synchronization.Tests
         private IFunctionMetadataService _functionMetadataService;
         private IExecutingService _concurrencyExecutingService;
         private ITxHub _txHub;
-        private IChainManagerBasic _chainManagerBasic;
+        private IChainManager _chainManager;
 
         private IBlockSynchronizer _blockSynchronizer;
 
@@ -49,7 +49,7 @@ namespace AElf.Synchronization.Tests
                     _functionMetadataService), _transactionTraceManager, _stateStore,
                 new ChainContextService(GetChainService()));
             _txHub = txHub;
-            _chainManagerBasic = new ChainManagerBasic(dataStore);
+            _chainManager = new ChainManager(dataStore);
         }
 
         public IBlockSynchronizer GetBlockSynchronizer()
@@ -101,7 +101,7 @@ namespace AElf.Synchronization.Tests
         public IBlockExecutor GetBlockExecutor()
         {
             return new BlockExecutor(GetChainService(), _concurrencyExecutingService, 
-                _transactionResultManager, null, null, _txHub, _chainManagerBasic,_stateStore);
+                _transactionResultManager, null, null, _txHub, _chainManager,_stateStore);
         }
     }
 }
